- removes the App folder / namespace - adds guarded = ['id'] to model - changes `web` middleware to `api` - adds type to props in structure migrations - updates helpers from classes to services
