--- src/App/Http/Controllers/Create.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace LaravelEnso\Menus\App\Http\Controllers;
-
-use Illuminate\Routing\Controller;
-use LaravelEnso\Menus\App\Forms\Builders\MenuForm;
-
-class Create extends Controller
-{
-    public function __invoke(MenuForm $form)
-    {
-        return ['form' => $form->create()];
-    }
-}
