--- src/App/Http/Controllers/Store.php
@@ -1,21 +1,0 @@
-<?php
-
-namespace LaravelEnso\Menus\App\Http\Controllers;
-
-use Illuminate\Routing\Controller;
-use LaravelEnso\Menus\App\Http\Requests\ValidateMenuRequest;
-use LaravelEnso\Menus\App\Models\Menu;
-
-class Store extends Controller
-{
-    public function __invoke(ValidateMenuRequest $request, Menu $menu)
-    {
-        $menu->fill($request->validated())->save();
-
-        return [
-            'message' => __('The menu was created!'),
-            'redirect' => 'system.menus.edit',
-            'param' => ['menu' => $menu->id],
-        ];
-    }
-}
