--- src/App/Http/Controllers/TableData.php
@@ -1,14 +1,0 @@
-<?php
-
-namespace LaravelEnso\Menus\App\Http\Controllers;
-
-use Illuminate\Routing\Controller;
-use LaravelEnso\Menus\App\Tables\Builders\MenuTable;
-use LaravelEnso\Tables\App\Traits\Data;
-
-class TableData extends Controller
-{
-    use Data;
-
-    protected $tableClass = MenuTable::class;
-}
