--- src/App/Http/Controllers/Update.php
@@ -1,17 +1,0 @@
-<?php
-
-namespace LaravelEnso\Menus\App\Http\Controllers;
-
-use Illuminate\Routing\Controller;
-use LaravelEnso\Menus\App\Http\Requests\ValidateMenuRequest;
-use LaravelEnso\Menus\App\Models\Menu;
-
-class Update extends Controller
-{
-    public function __invoke(ValidateMenuRequest $request, Menu $menu)
-    {
-        $menu->update($request->validated());
-
-        return ['message' => __('The menu was successfully updated')];
-    }
-}
