--- src/App/Services/Organizer.php
@@ -1,36 +1,0 @@
-<?php
-
-namespace LaravelEnso\Menus\App\Services;
-
-use LaravelEnso\Helpers\App\Classes\Obj;
-use LaravelEnso\Menus\App\Models\Menu;
-
-class Organizer
-{
-    private Obj $menus;
-
-    public function __construct(array $menus)
-    {
-        $this->menus = new Obj($menus);
-    }
-
-    public function handle(): void
-    {
-        $this->organize($this->menus);
-    }
-
-    private function organize(Obj $menus): void
-    {
-        $menus->each(fn ($menu, $index) => $this->reorder($menu, $index));
-    }
-
-    private function reorder(Obj $menu, int $index): void
-    {
-        Menu::find($menu->get('id'))
-            ->update(['order_index' => ($index + 1) * 10]);
-
-        if ($menu->get('has_children')) {
-            $this->organize($menu->get('children'));
-        }
-    }
-}
