+++ src/Http/Controllers/InitTable.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace LaravelEnso\Menus\Http\Controllers;
+
+use Illuminate\Routing\Controller;
+use LaravelEnso\Menus\Tables\Builders\MenuTable;
+use LaravelEnso\Tables\Traits\Init;
+
+class InitTable extends Controller
+{
+    use Init;
+
+    protected $tableClass = MenuTable::class;
+}
