+++ src/Http/Controllers/Store.php
@@ -1,0 +1,21 @@
+<?php
+
+namespace LaravelEnso\Menus\Http\Controllers;
+
+use Illuminate\Routing\Controller;
+use LaravelEnso\Menus\Http\Requests\ValidateMenuRequest;
+use LaravelEnso\Menus\Models\Menu;
+
+class Store extends Controller
+{
+    public function __invoke(ValidateMenuRequest $request, Menu $menu)
+    {
+        $menu->fill($request->validated())->save();
+
+        return [
+            'message' => __('The menu was created!'),
+            'redirect' => 'system.menus.edit',
+            'param' => ['menu' => $menu->id],
+        ];
+    }
+}
